Work with RIDL datasets (datasets)

Work with RIDL datasets (datasets)

Usage

dataset_create(metadata)

dataset_show(id)

dataset_update(id, metadata)

dataset_patch(id, metadata)

dataset_delete(id)

Arguments

metadata

Metadata created by dataset_metadata().

id

The id or name of the dataset.

Value

The dataset.

Details

You must have the necessary permissions to create, edit, or delete datasets.

Note that several fields are required for dataset_create() and dataset_update() operations to succeed. Consult dataset_metadata() for the details.

For dataset_update()/dataset_patch() operations, it is recommended to call dataset_show(), make the desired changes to the result, and then call dataset_update()/dataset_patch() with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the metadata.

Examples

#-----
# test search in prod
Sys.unsetenv("USE_UAT")
riddle::dataset_show("unhcr-cbi-americas-quarterly-report")
#> FalseNULLNULL74143486-95d1-44bb-a1e8-b168a9b36a22cawiUnited Nations High Commissioner for Refugeesyes2022-02-142021-04-07not_availableRegionalUNSPECIFIEDd6da9482-9062-46c7-a3ef-c47bf71937ccpersonally_identifiableFALSE27a3YPSk4nL6ARa4xGLAhhyuNULLNULLNULLNULL2022-02-16T21:36:45.6817262022-02-16T21:37:52.927952unhcr-cbi-americas-quarterly-reportDataset imported from KoBoToolbox
#> KoBoToolbox owner: cbi_americas
#> Sector: Ayuda Humanitaria - Coordinación / Gestión de la Información
#> Country: Panama (PAN)40programme_monitoringlist(id = "a402d93c-edc4-4885-95a3-7463c433f50a", name = "americas-regional-dataset", title = "Americas : Regional Datasets", type = "data-container", description = "", image_url = "", created = "2020-06-11T16:46:16.394107", is_organization = TRUE, approval_status = "approved", state = "active")a3YPSk4nL6ARa4xGLAhhyua402d93c-edc4-4885-95a3-7463c433f50aFALSErawlist()activeUNHCR CBI Americas: Quarterly ReportdatasetcountryNULLrestrictedlist(cache_last_updated = c(NA, NA, NA, NA), cache_url = c(NA, NA, NA, NA), created = c("2022-02-16T21:36:46.648492", "2022-02-16T21:36:48.150908", "2022-02-16T21:36:49.537820", "2022-02-16T21:36:51.007008"), datastore_active = c(FALSE, FALSE, FALSE, FALSE), description = c("Questionnaire imported from the KoBo survey", "JSON data imported from the KoBoToolbox survey. ", "CSV data imported from the KoBoToolbox survey. ", "XLS data imported from the KoBoToolbox survey. "), file_type = c("questionnaire", 
#> "microdata", "microdata", "microdata"), format = c("XLS", "JSON", "CSV", "XLS"), hash = c("", "", "", ""), id = c("6ec5d932-fd5b-41c2-aa7c-7f86096abf86", "44ac8836-dacd-46c0-8a46-97f778522b2b", "0d5e31d4-c052-442e-82a2-13707571a2e2", "d5d9c8e2-4a5b-4c08-a851-3409c4b88c3b"), kobo_details = list(kobo_asset_id = c("a3YPSk4nL6ARa4xGLAhhyu", "a3YPSk4nL6ARa4xGLAhhyu", "a3YPSk4nL6ARa4xGLAhhyu", "a3YPSk4nL6ARa4xGLAhhyu"), kobo_download_status = c("complete", "complete", "complete", "complete"), kobo_download_attempts = c(1, 
#> 1, 1, 1), kobo_last_updated = c("2022-02-16T21:36:59.955251", "2022-02-16T21:37:13.109264", "2022-02-16T21:37:25.613627", "2022-02-16T21:37:43.018402"), kobo_export_id = c(NA, NA, "e4xaPtuZLFu56nabdmawxn", "eumdDkexmhJgY7VJUYcXcw"), kobo_submission_count = c(NA, 94, 94, 94)), kobo_type = c("questionnaire", "data", "data", "data"), last_modified = c("2022-02-16T21:37:02.522322", "2022-02-16T21:37:14.549043", "2022-02-16T21:37:27.141097", "2022-02-16T21:37:45.654386"), metadata_modified = c("2022-02-16T21:37:02.843844", 
#> "2022-02-16T21:37:14.845454", "2022-02-16T21:37:27.447127", "2022-02-16T21:37:46.406802"), mimetype = c("application/x-empty", "application/x-empty", "application/x-empty", "application/x-empty"), mimetype_inner = c(NA, NA, NA, NA), name = c("Questionnaire XLS", "Survey json data", "Survey csv data", "Survey xls data"), package_id = c("d6da9482-9062-46c7-a3ef-c47bf71937cc", "d6da9482-9062-46c7-a3ef-c47bf71937cc", "d6da9482-9062-46c7-a3ef-c47bf71937cc", "d6da9482-9062-46c7-a3ef-c47bf71937cc"), position = 0:3, 
#>     resource_type = c(NA, NA, NA, NA), size = c(0, 0, 0, 0), state = c("active", "active", "active", "active"), type = c("attachment", "data", "data", "data"), url = c("https://ridl.unhcr.org/dataset/d6da9482-9062-46c7-a3ef-c47bf71937cc/resource/6ec5d932-fd5b-41c2-aa7c-7f86096abf86/download/a3ypsk4nl6ara4xglahhyu.xls", "https://ridl.unhcr.org/dataset/d6da9482-9062-46c7-a3ef-c47bf71937cc/resource/44ac8836-dacd-46c0-8a46-97f778522b2b/download/a3ypsk4nl6ara4xglahhyu_data.json", "https://ridl.unhcr.org/dataset/d6da9482-9062-46c7-a3ef-c47bf71937cc/resource/0d5e31d4-c052-442e-82a2-13707571a2e2/download/a3ypsk4nl6ara4xglahhyu_data.csv", 
#>     "https://ridl.unhcr.org/dataset/d6da9482-9062-46c7-a3ef-c47bf71937cc/resource/d5d9c8e2-4a5b-4c08-a851-3409c4b88c3b/download/a3ypsk4nl6ara4xglahhyu_data.xls"), url_type = c("upload", "upload", "upload", "upload"), visibility = c("public", "restricted", "restricted", "restricted"), date_range_end = c(NA, "2022-02-14", "2022-02-14", "2022-02-14"), date_range_start = c(NA, "2021-04-07", "2021-04-07", "2021-04-07"), identifiability = c(NA, "personally_identifiable", "personally_identifiable", "personally_identifiable"
#>     ), process_status = c(NA, "raw", "raw", "raw"), version = c(NA, "1", "1", "1"))list()list()list()list()
#> # A tibble: 1 × 54
#>   admin…¹ archi…² author autho…³ clean…⁴ creat…⁵ data_…⁶ data_…⁷ data_…⁸ data_…⁹
#>   <chr>   <chr>   <lgl>  <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 ""      False   NA     NA      ""      741434… ""      ""      cawi    United…
#> # … with 44 more variables: data_sensitivity <chr>, date_range_end <chr>,
#> #   date_range_start <chr>, external_access_level <chr>, geog_coverage <chr>,
#> #   geographies <chr>, id <chr>, identifiability <chr>, isopen <lgl>,
#> #   keywords <chr>, kobo_asset_id <chr>, license_id <lgl>, license_title <lgl>,
#> #   maintainer <lgl>, maintainer_email <lgl>, metadata_created <chr>,
#> #   metadata_modified <chr>, name <chr>, notes <chr>, num_resources <int>,
#> #   num_tags <int>, operational_purpose_of_data <chr>, organization <list>, …

p <- riddle::dataset_show('rms_v4')
#> The RMS survey was integrated with the HFS Protection Monitoring module to streamline collection. Therefore, some changes were made to the original RMS form.FalseNULLNULL5cdbcb1d-4cdb-48e6-9c4d-2f3a50d2c9c0triUnited Nations High Commissioner for Refugeesyes2022-07-302022-07-11not_availableGTM8c7eb4c1-b40e-4fc5-abae-896655f64cb2personally_identifiableFALSEc("3", "5", "15", "62")aQTbMZSPH4vwMJTtk4xjHTNULLNULLNULLNULL2022-08-31T22:49:50.6941752022-12-08T15:55:38.364640rms_v4RMS: Survey to collect data for core impact and core outcome indicators in a standardized way.  Protection Monitoring: collection of information for the analysis of profiles with protection needs.  Interviews conducted by phone by a specialized UNHCR provider.
#> Observations: 3981.  Latest submission: July 25, 2022.53protection_monitoringlist(id = "99f73637-5da9-4199-bade-1fab89343438", name = "guatemala", title = "Guatemala", type = "data-container", description = "", image_url = "", created = "2018-07-12T21:00:37.554859", is_organization = TRUE, approval_status = "approved", state = "active")aQTbMZSPH4vwMJTtk4xjHT99f73637-5da9-4199-bade-1fab89343438FALSErawnonprobability_availabilityResults Monitoring Survey (RMS) and Protection MonitoringactiveResults Monitoring Survey (RMS) and Protection MonitoringdatasetindividualNULLrestrictedlist(cache_last_updated = c(NA, NA, NA, NA, NA), cache_url = c(NA, NA, NA, NA, NA), created = c("2022-08-31T22:49:51.777678", "2022-08-31T22:49:55.450262", "2022-08-31T22:49:57.089230", "2022-10-03T14:32:11.734505", "2022-10-04T13:59:30.533323"), datastore_active = c(FALSE, FALSE, FALSE, FALSE, FALSE), description = c("Questionnaire imported from the KoBo survey", "CSV data imported from the KoBoToolbox survey. ", "XLS data imported from the KoBoToolbox survey. ", "Data from Kobo in XML format with values and headers.  3 October 2022.", 
#> ""), file_type = c("questionnaire", "microdata", "microdata", "microdata", "report"), format = c("XLS", "CSV", "XLS", "XLSX", "HTML"), hash = c("", "", "", "", ""), id = c("2c6a805f-77d4-4b3d-9873-5480d9943472", "491c3718-a4e6-40c8-8289-90199e5457cb", "92ae1611-149c-4e4a-b83a-eaeb25942309", "21b3ce9f-5653-4819-a6ac-df19283691d1", "29d4bdba-b599-4e9e-8f17-eadbe6fadcc1"), kobo_details = list(kobo_asset_id = c("aQTbMZSPH4vwMJTtk4xjHT", "aQTbMZSPH4vwMJTtk4xjHT", "aQTbMZSPH4vwMJTtk4xjHT", NA, NA), kobo_download_status = c("complete", 
#> "complete", "error", NA, NA), kobo_download_attempts = c(1, 1, 5, NA, NA), kobo_last_updated = c("2022-08-31T22:50:04.948178", "2022-08-31T22:50:18.546376", "2022-12-08T15:55:37.770805", NA, NA), kobo_export_id = c(NA, "eBoP9fbADL6mBY4JCsdxui", "eJFS5SeMNLYVQKxvnEKwBQ", NA, NA), kobo_filter_fields_from_all_versions = c(NA, TRUE, TRUE, NA, NA), kobo_filter_group_sep = c(NA, "/", "/", NA, NA), kobo_filter_fields = list(NULL, list(), list(), NULL, NULL), kobo_filter_query = c(NA, NA, NA, NA, NA), kobo_filter_multiple_select = c(NA, 
#> "both", "both", NA, NA), kobo_filter_hierarchy_in_labels = c(NA, TRUE, TRUE, NA, NA), kobo_filter_language = c(NA, "_default", "_default", NA, NA), kobo_last_error_response = c(NA, NA, "Failed to download KoboToolbox data resource 92ae1611-149c-4e4a-b83a-eaeb25942309. Last export status from Kobo: {'url': 'https://kobo.unhcr.org/api/v2/assets/aQTbMZSPH4vwMJTtk4xjHT/exports/eJFS5SeMNLYVQKxvnEKwBQ/?format=json', 'status': 'processing', 'messages': {}, 'uid': 'eJFS5SeMNLYVQKxvnEKwBQ', 'date_created': '2022-12-08T15:47:56.455119Z', 'last_submission_time': None, 'result': None, 'data': {'lang': '_default', 'name': None, 'type': 'xls', 'source': 'https://kobo.unhcr.org/api/v2/assets/aQTbMZSPH4vwMJTtk4xjHT/?format=json', 'flatten': True, 'group_sep': '/', 'multiple_select': 'both', 'hierarchy_in_labels': True, 'fields_from_all_versions': True}}", 
#> NA, NA)), kobo_type = c("questionnaire", "data", "data", "", ""), last_modified = c("2022-08-31T22:50:07.684601", "2022-08-31T22:50:20.090377", "2022-08-31T22:50:33.070115", "2022-10-03T14:32:11.536587", "2022-10-04T13:59:30.321887"), metadata_modified = c("2022-08-31T22:50:08.260981", "2022-08-31T22:50:20.447250", "2022-12-08T15:55:38.370890", "2022-10-03T14:32:18.063693", "2022-10-04T13:59:36.012003"), mimetype = c("application/x-empty", "application/x-empty", "application/x-empty", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", 
#> "text/html"), mimetype_inner = c(NA, NA, NA, NA, NA), name = c("Questionnaire XLS", "Survey csv data", "Survey xls data", "RMS XML values and headers", "Crunching report"), package_id = c("8c7eb4c1-b40e-4fc5-abae-896655f64cb2", "8c7eb4c1-b40e-4fc5-abae-896655f64cb2", "8c7eb4c1-b40e-4fc5-abae-896655f64cb2", "8c7eb4c1-b40e-4fc5-abae-896655f64cb2", "8c7eb4c1-b40e-4fc5-abae-896655f64cb2"), position = 0:4, resource_type = c(NA, NA, NA, NA, NA), size = c(0, 0, 0, 1967301, 56601325), state = c("active", 
#> "active", "active", "active", "active"), type = c("attachment", "data", "data", "data", "attachment"), url = c("https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/2c6a805f-77d4-4b3d-9873-5480d9943472/download/aqtbmzsph4vwmjttk4xjht.xls", "https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/491c3718-a4e6-40c8-8289-90199e5457cb/download/survey-csv-data__aqtbmzsph4vwmjttk4xjht_data.csv", "https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/92ae1611-149c-4e4a-b83a-eaeb25942309/download/survey-xls-data__aqtbmzsph4vwmjttk4xjht_data.xls", 
#> "https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/21b3ce9f-5653-4819-a6ac-df19283691d1/download/rms_v4_xml_values_and_headers.xlsx", "https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/29d4bdba-b599-4e9e-8f17-eadbe6fadcc1/download/exploration.html"), url_type = c("upload", "upload", "upload", "upload", "upload"), visibility = c("public", "restricted", "restricted", "restricted", "public"), date_range_end = c(NA, "2022-07-30", "2022-07-30", "2022-07-30", 
#> NA), date_range_start = c(NA, "2022-07-11", "2022-07-11", "2022-07-11", NA), identifiability = c(NA, "personally_identifiable", "personally_identifiable", "personally_identifiable", NA), process_status = c(NA, "raw", "raw", "raw", NA), version = c(NA, "1", "1", "1", NA), original_url = c(NA, NA, NA, "", ""))list(display_name = c("Monitoring", "RMS", "protection"), id = c("c58abf44-15dd-442b-a712-a814c8a04c1c", "2f0595e3-3ac5-4c85-8deb-e931cf390a40", "58aa88bd-f6b9-49cc-843b-b2c6d6b8f1c2"), name = c("Monitoring", "RMS", "protection"), state = c("active", "active", "active"), vocabulary_id = c(NA, NA, NA))list()list()list()
list_of_ressources <- p[["resources"]][[1]]
list_of_ressources
#> # A tibble: 5 × 31
#>   cache_las…¹ cache…² created datas…³ descr…⁴ file_…⁵ format hash  id    kobo_…⁶
#>   <lgl>       <lgl>   <chr>   <lgl>   <chr>   <chr>   <chr>  <chr> <chr> <chr>  
#> 1 NA          NA      2022-0… FALSE   "Quest… questi… XLS    ""    2c6a… aQTbMZ…
#> 2 NA          NA      2022-0… FALSE   "CSV d… microd… CSV    ""    491c… aQTbMZ…
#> 3 NA          NA      2022-0… FALSE   "XLS d… microd… XLS    ""    92ae… aQTbMZ…
#> 4 NA          NA      2022-1… FALSE   "Data … microd… XLSX   ""    21b3… NA     
#> 5 NA          NA      2022-1… FALSE   ""      report  HTML   ""    29d4… NA     
#> # … with 33 more variables: kobo_details$kobo_download_status <chr>,
#> #   $kobo_download_attempts <int>, $kobo_last_updated <chr>,
#> #   $kobo_export_id <chr>, $kobo_filter_fields_from_all_versions <lgl>,
#> #   $kobo_filter_group_sep <chr>, $kobo_filter_fields <list>,
#> #   $kobo_filter_query <lgl>, $kobo_filter_multiple_select <chr>,
#> #   $kobo_filter_hierarchy_in_labels <lgl>, $kobo_filter_language <chr>,
#> #   $kobo_last_error_response <chr>, kobo_type <chr>, last_modified <chr>, …

#-----
# Test create in UAT
Sys.setenv(USE_UAT=1)

# p <- riddle::dataset_show('testedouard2')

### TO FIX -- geographies: Missing value
## __type: Validation Error ---
## cf https://github.com/okfn/ckanext-unhcr/blob/master/ckanext/unhcr/schemas/dataset.json#L670:L682

m <- riddle::dataset_metadata(title = "Motor Trend Car Road Test two",
                      name = "mtcars_ed",
                      notes = "The data was extracted from the 1974 Motor Trend
                      US magazine, and comprises fuel consumption and 10 aspects
                      of automobile design and performance for 32 automobiles
                      (1973–74 models).",
                      owner_org = "Americas",
                      visibility = "public",
                      geographies = "UNSPECIFIED",
                      external_access_level = "open_access",
                      data_collector = "Motor Trend",
                      keywords = keywords[c("Environment", "Other")],
                      unit_of_measurement = "car",
                      data_collection_technique = "oth",
                      archived = "False")
# ## For the above to work - you need to make sure you have at least editor access
# to the corresponding container - i.e. owner_org = "exercise-container"
#p <- riddle::dataset_create(metadata = m)
# The return value is a representation of the dataset we just created in
# RIDL that you could inspect like any other R object.
#p

#-----
# Test create in prod
Sys.unsetenv("USE_UAT")
# m1 <- riddle::dataset_metadata(title = "Test",
#                       name = "Test",
#                       notes = "The data was extracted from kobo.",
#                       owner_org = "Americas : Regional Datasets",
#                       visibility = "public",
#                       geographies = "UNSPECIFIED",
#                       external_access_level = "open_access",
#                       data_collector = "UNHCR",
#                       keywords = keywords[c("Environment", "Other")],
#                       unit_of_measurement = "car",
#                       data_collection_technique = "oth",
#                       archived = "False")
# p <- riddle::dataset_create(metadata = m1)