Work with RIDL resources (files)

Work with RIDL resources (files)

Usage

resource_create(pkgid, res_metadata)

resource_update(id, res_metadata)

resource_patch(id, res_metadata)

resource_delete(id)

resource_create(pkgid, res_metadata)

resource_update(id, res_metadata)

resource_patch(id, res_metadata)

resource_delete(id)

Arguments

res_metadata

Metadata created by resource_metadata().

id

The id or name of the resource.

dataset_id

The id or name of the dataset to which this resource belongs to.

Value

The resource.

Details

You must have the necessary permissions to create, edit, or delete datasets and their resources.

Note that several fields are required for resource_create() and resource_update() operations to succeed. Consult resource_metadata() for the details.

For resource_update()/resource_patch() operations, it is recommended to call resource_show(), make the desired changes to the result, and then call resource_update()/resource_patch() with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the metadata.

Examples

## let's get again the details of the dataset we want to add the resource in..
#-----
# Test search in prod
Sys.unsetenv("USE_UAT")
p <- riddle::dataset_search("rms_v4")
p
#> # A tibble: 10 × 54
#>    admin_notes    archi…¹ author autho…² clean…³ creat…⁴ data_…⁵ data_…⁶ data_…⁷
#>    <chr>          <chr>   <lgl>  <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 "The RMS surv… False   NA     NA      ""      5cdbcb… ""      ""      tri    
#>  2 "R4V - RMRP B… True    NA     NA      ""      3b109e… "Restr… "Tool:… cawi   
#>  3 ""             False   NA     NA      ""      183804… ""      ""      tri    
#>  4 "Uploaded dat… False   NA     NA      "Uploa… 10a321… "Data … "Data … cami   
#>  5 ""             False   NA     NA      ""      de99be… ""      ""      tri    
#>  6 ""             True    NA     NA      ""      1b633b… ""      ""      tri    
#>  7  NA            False   NA     NA       NA     b5d85f…  NA      NA     tri    
#>  8  NA            False   NA     NA       NA     b5d85f…  NA      NA     tri    
#>  9 ""             False   NA     NA      ""      e44763… ""      ""      f2f    
#> 10 ""             False   NA     NA      ""      9a1c9d… ""      ""      capi   
#> # … with 45 more variables: data_collector <chr>, data_sensitivity <chr>,
#> #   date_range_end <chr>, date_range_start <chr>, external_access_level <chr>,
#> #   geog_coverage <chr>, geographies <list>, id <chr>, identifiability <chr>,
#> #   isopen <lgl>, keywords <list>, kobo_asset_id <chr>, license_id <lgl>,
#> #   license_title <lgl>, maintainer <lgl>, maintainer_email <lgl>,
#> #   metadata_created <chr>, metadata_modified <chr>, name <chr>, notes <chr>,
#> #   num_resources <int>, num_tags <int>, operational_purpose_of_data <list>, …

#-----
# Test search in uat
Sys.setenv(USE_UAT=1)
p <- riddle::dataset_search("tests")
p
#> # A tibble: 10 × 54
#>    admin_notes    archi…¹ author autho…² clean…³ creat…⁴ data_…⁵ data_…⁶ data_…⁷
#>    <chr>          <chr>   <lgl>  <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ""             False   NA     NA      ""      daf9d9… ""      ""      nf     
#>  2 ""             True    NA     NA      ""      1b633b… ""      ""      tri    
#>  3 ""             False   NA     NA      ""      9a7869… ""      ""      tri    
#>  4 ""             False   NA     NA      ""      a87b60… ""      ""      f2f    
#>  5 ""             False   NA     NA      ""      f1b4d3… ""      ""      tri    
#>  6 "The dataset … False   NA     NA      "For p… 274891… ""      "The m… capi   
#>  7 ""             False   NA     NA      ""      aaf6de… ""      ""      tri    
#>  8 ""             True    NA     NA      "The i… cf0c93… "The K… "Eleve… tri    
#>  9 ""             True    NA     NA      "The i… cf0c93… "The K… "Six U… tri    
#> 10 ""             False   NA     NA      ""      f1b4d3… ""      ""      f2f    
#> # … with 45 more variables: data_collector <chr>, data_sensitivity <chr>,
#> #   date_range_end <chr>, date_range_start <chr>, external_access_level <chr>,
#> #   geog_coverage <chr>, geographies <list>, id <chr>, identifiability <chr>,
#> #   isopen <lgl>, keywords <list>, kobo_asset_id <chr>, license_id <lgl>,
#> #   license_title <lgl>, maintainer <lgl>, maintainer_email <lgl>,
#> #   metadata_created <chr>, metadata_modified <chr>, name <chr>, notes <chr>,
#> #   num_resources <int>, num_tags <int>, operational_purpose_of_data <list>, …
##take the first one
ridlid <- as.character(p[9, c("id")])

## let's get again the details of the dataset we want to add the resource in..
#r <- riddle::resource_create(dataset_id = ridlid, res_metadata = m)
# Like before, the return value is a tibble representation of the resource.
#r
 
## and now can search for it - checking it is correctly there... 
##riddle::resource_search("name:mtcarsriddle")

# And once we’re done experimenting with the API, we should take down our
# toy dataset since we don’t really need it on RIDL.
#riddle::dataset_delete(p$id)

#-----
# Test search in prod
Sys.unsetenv("USE_UAT")
p <- riddle::dataset_show('rms_v4')
dataset_id_rms <- p$id
list_of_ressources <- p[["resources"]][[1]]
knitr::kable(list_of_ressources)
#> Warning: provided 13 variables to replace 1 variables
#> 
#> 
#> |cache_last_updated |cache_url |created                    |datastore_active |description                                                            |file_type     |format |hash |id                                   |kobo_details           |kobo_type     |last_modified              |metadata_modified          |mimetype                                                          |mimetype_inner |name                       |package_id                           | position|resource_type |     size|state  |type       |url                                                                                                                                                                         |url_type |visibility |date_range_end |date_range_start |identifiability         |process_status |version |original_url |
#> |:------------------|:---------|:--------------------------|:----------------|:----------------------------------------------------------------------|:-------------|:------|:----|:------------------------------------|:----------------------|:-------------|:--------------------------|:--------------------------|:-----------------------------------------------------------------|:--------------|:--------------------------|:------------------------------------|--------:|:-------------|--------:|:------|:----------|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:--------|:----------|:--------------|:----------------|:-----------------------|:--------------|:-------|:------------|
#> |NA                 |NA        |2022-08-31T22:49:51.777678 |FALSE            |Questionnaire imported from the KoBo survey                            |questionnaire |XLS    |     |2c6a805f-77d4-4b3d-9873-5480d9943472 |aQTbMZSPH4vwMJTtk4xjHT |questionnaire |2022-08-31T22:50:07.684601 |2022-08-31T22:50:08.260981 |application/x-empty                                               |NA             |Questionnaire XLS          |8c7eb4c1-b40e-4fc5-abae-896655f64cb2 |        0|NA            |        0|active |attachment |https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/2c6a805f-77d4-4b3d-9873-5480d9943472/download/aqtbmzsph4vwmjttk4xjht.xls                       |upload   |public     |NA             |NA               |NA                      |NA             |NA      |NA           |
#> |NA                 |NA        |2022-08-31T22:49:55.450262 |FALSE            |CSV data imported from the KoBoToolbox survey.                         |microdata     |CSV    |     |491c3718-a4e6-40c8-8289-90199e5457cb |aQTbMZSPH4vwMJTtk4xjHT |data          |2022-08-31T22:50:20.090377 |2022-08-31T22:50:20.447250 |application/x-empty                                               |NA             |Survey csv data            |8c7eb4c1-b40e-4fc5-abae-896655f64cb2 |        1|NA            |        0|active |data       |https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/491c3718-a4e6-40c8-8289-90199e5457cb/download/survey-csv-data__aqtbmzsph4vwmjttk4xjht_data.csv |upload   |restricted |2022-07-30     |2022-07-11       |personally_identifiable |raw            |1       |NA           |
#> |NA                 |NA        |2022-08-31T22:49:57.089230 |FALSE            |XLS data imported from the KoBoToolbox survey.                         |microdata     |XLS    |     |92ae1611-149c-4e4a-b83a-eaeb25942309 |aQTbMZSPH4vwMJTtk4xjHT |data          |2022-08-31T22:50:33.070115 |2022-12-08T15:55:38.370890 |application/x-empty                                               |NA             |Survey xls data            |8c7eb4c1-b40e-4fc5-abae-896655f64cb2 |        2|NA            |        0|active |data       |https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/92ae1611-149c-4e4a-b83a-eaeb25942309/download/survey-xls-data__aqtbmzsph4vwmjttk4xjht_data.xls |upload   |restricted |2022-07-30     |2022-07-11       |personally_identifiable |raw            |1       |NA           |
#> |NA                 |NA        |2022-10-03T14:32:11.734505 |FALSE            |Data from Kobo in XML format with values and headers.  3 October 2022. |microdata     |XLSX   |     |21b3ce9f-5653-4819-a6ac-df19283691d1 |NA                     |              |2022-10-03T14:32:11.536587 |2022-10-03T14:32:18.063693 |application/vnd.openxmlformats-officedocument.spreadsheetml.sheet |NA             |RMS XML values and headers |8c7eb4c1-b40e-4fc5-abae-896655f64cb2 |        3|NA            |  1967301|active |data       |https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/21b3ce9f-5653-4819-a6ac-df19283691d1/download/rms_v4_xml_values_and_headers.xlsx               |upload   |restricted |2022-07-30     |2022-07-11       |personally_identifiable |raw            |1       |             |
#> |NA                 |NA        |2022-10-04T13:59:30.533323 |FALSE            |                                                                       |report        |HTML   |     |29d4bdba-b599-4e9e-8f17-eadbe6fadcc1 |NA                     |              |2022-10-04T13:59:30.321887 |2022-10-04T13:59:36.012003 |text/html                                                         |NA             |Crunching report           |8c7eb4c1-b40e-4fc5-abae-896655f64cb2 |        4|NA            | 56601325|active |attachment |https://ridl.unhcr.org/dataset/8c7eb4c1-b40e-4fc5-abae-896655f64cb2/resource/29d4bdba-b599-4e9e-8f17-eadbe6fadcc1/download/exploration.html                                 |upload   |public     |NA             |NA               |NA                      |NA             |NA      |             |

new_attachment <- riddle::resource_metadata(type = "attachment",
                       url = "resource.R", 
# upload = httr::upload_file(here::here("R","resource.R") ),
## # Error:
# ! All columns in a tibble must be vectors.
# ✖ Column `upload` is a `form_file` object.
                       name = "Rscript",
                       format = "R",
                       file_type = "report",
                       version = "1",
                       visibility = "public" )

## Now testing adding the file "resource.R" as an attachment in the 
## dataset 'rms_v4'

# but not working.... 
# resource_create(dataset_id = dataset_id_rms, 
#                         res_metadata = new_attachment )
# Error in ridl(action = "resource_create", dataset_id = dataset_id, !!!res_metadata, : 
# __type: Not Found Error
# message: Not found: No dataset id provided, cannot check auth.

m <- riddle::resource_metadata(type = "data",
                       url = "mtcars.csv",
#  upload = httr::upload_file(system.file("extdata/mtcars.csv", package = "readr")),         
## # Error:
# ! All columns in a tibble must be vectors.
# ✖ Column `upload` is a `form_file` object.
                       name = "mtcars.csv",
                       format = "csv",
                       file_type = "microdata",
                       date_range_start = "1973-01-01",
                       date_range_end = "1973-12-31",
                       version = "1",
                       visibility = "public",
                       process_status = "raw",
                       identifiability = "anonymized_public")
# same pb... 
# resource_create(dataset_id = dataset_id_rms, 
#                          res_metadata = m )
# Error in ridl(action = "resource_create", dataset_id = dataset_id, !!!res_metadata, : 
# __type: Not Found Error
# message: Not found: No dataset id provided, cannot check auth.