Searches for datasets and resources satisfying a given criteria.

Work with RIDL resources (files)

Usage

dataset_search(q = NULL, rows = NULL, start = NULL)

resource_search(query = NULL, rows = NULL, start = NULL)

resource_search(query = NULL, rows = NULL, start = NULL)

Arguments

q, query

The search query.

rows

The maximum number of matching rows (datasets) to return. (optional, default: 10, upper limit: 1000)

start

The offset in the complete result for where the set of returned datasets should begin.

res_metadata

Metadata created by resource_metadata().

id

The id or name of the resource.

pkgid

The id or name of the dataset to which this resource belongs to.

Value

A tibble with the search results.

The resource.

Details

You must have the necessary permissions to create, edit, or delete datasets and their resources.

Note that several fields are required for resource_create() and resource_update() operations to succeed. Consult resource_metadata() for the details.

For resource_update()/resource_patch() operations, it is recommended to call resource_show(), make the desired changes to the result, and then call resource_update()/resource_patch() with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the metadata.

Examples

#-----
# Test search in prod
Sys.unsetenv("USE_UAT")
p <- riddle::dataset_search(q = "cbi")
p
#> # A tibble: 10 × 54
#>    admin_notes archived author author_…¹ clean…² creat…³ data_…⁴ data_…⁵ data_…⁶
#>    <chr>       <chr>    <lgl>  <lgl>     <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ""          False    NA     NA        ""      c05196… ""      ""      tri    
#>  2 ""          True     NA     NA        "The i… cf0c93… "The K… "Eleve… tri    
#>  3 ""          True     NA     NA        "The i… cf0c93… "The K… "Six U… tri    
#>  4 ""          False    NA     NA        ""      823101… ""      ""      tri    
#>  5 ""          True     NA     NA        ""      466fee… ""      ""      tri    
#>  6 ""          False    NA     NA        ""      b5d85f… ""      ""      tri    
#>  7 ""          False    NA     NA        ""      565802… ""      ""      tri    
#>  8 NA          False    NA     NA         NA     b5d85f…  NA      NA     tri    
#>  9 NA          False    NA     NA         NA     b5d85f…  NA      NA     tri    
#> 10 NA          False    NA     NA         NA     b5d85f…  NA      NA     tri    
#> # … with 45 more variables: data_collector <chr>, date_range_end <chr>,
#> #   date_range_start <chr>, external_access_level <chr>, geog_coverage <chr>,
#> #   geographies <list>, id <chr>, identifiability <chr>, isopen <lgl>,
#> #   keywords <list>, kobo_asset_id <chr>, license_id <lgl>,
#> #   license_title <lgl>, maintainer <lgl>, maintainer_email <lgl>,
#> #   metadata_created <chr>, metadata_modified <chr>, name <chr>, notes <chr>,
#> #   num_resources <int>, num_tags <int>, operational_purpose_of_data <list>, …
p$id
#>  [1] "222379a3-1eb2-47da-8b16-cae6a9f579aa"
#>  [2] "34d55d82-3785-45bc-8857-08ac99e47917"
#>  [3] "c6a5e695-f300-47ce-90c3-a46f354485d9"
#>  [4] "cfe6085c-43bb-4e31-986d-2343f01d962f"
#>  [5] "f050bfc8-db6d-4a8f-98fb-f6532f4c6c56"
#>  [6] "81d83eaa-0631-4103-9697-548daa16aca4"
#>  [7] "bef4e75b-bc98-4759-8fa5-0b865b8f9624"
#>  [8] "8d681095-03fe-498c-8389-fd0380638fb1"
#>  [9] "9abe86f5-7048-4551-b6c1-8837112fc2a8"
#> [10] "cd6ee743-6e85-4deb-a40d-725f93d3894b"
## let's get again the details of the dataset we want to add the resource in..

Sys.unsetenv("USE_UAT")
p <- riddle::dataset_search("rms_v4")

Sys.setenv(USE_UAT=1)
p <- riddle::dataset_search("tests")

ridlid <- as.character(p[1, c("id")])

m <- riddle::resource_metadata(type = "data",
                       url = "mtcars.csv",
                       name = "mtcarsriddle",
                       format = "csv",
                       file_type = "microdata",
                       date_range_start = "1973-01-01",
                       date_range_end = "1973-12-31",
                       version = "1",
                       visibility = "public",
                       process_status = "raw",
                       identifiability = "anonymized_public",
upload = httr::upload_file(system.file("extdata/mtcars.csv", package = "readr")))

## let's get again the details of the dataset we want to add the resource in..
#r <- riddle::resource_create(ridlid, res_metadata = m)
# Like before, the return value is a tibble representation of the resource.
#r
 
## and now can search for it - checking it is correctly there... 
##riddle::resource_search("name:mtcarsriddle")

# And once we’re done experimenting with the API, we should take down our
# toy dataset since we don’t really need it on RIDL.
#riddle::dataset_delete(p$id)