Work with RIDL resources (files)

Work with RIDL resources (files)

Usage

resource_create(pkgid, metadata)

resource_update(id, metadata)

resource_patch(id, metadata)

resource_delete(id)

Arguments

pkgid

The id or name of the dataset to which this resource belongs to.

metadata

Metadata created by resource_metadata().

id

The id or name of the resource.

Value

The resource.

Details

You must have the necessary permissions to create, edit, or delete datasets and their resources.

Note that several fields are required for resource_create() and resource_update() operations to succeed. Consult resource_metadata() for the details.

For resource_update()/resource_patch() operations, it is recommended to call resource_show(), make the desired changes to the result, and then call resource_update()/resource_patch() with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the metadata.

Examples

library(riddle)
Sys.setenv(USE_UAT=1)
m <- dataset_metadata(title = "Motor Trend Car Road Tests",
                      name = "mtcars",
                      notes = "The data was extracted from the 1974 Motor Trend 
                      US magazine, and comprises fuel consumption and 10 aspects
                      of automobile design and performance for 32 automobiles 
                      (1973–74 models).",
                      owner_org = "exercise-container",
                      visibility = "public",
                      external_access_level = "open_access",
                      data_collector = "Motor Trend",
                      keywords = keywords[c("Environment", "Other")],
                      unit_of_measurement = "car",
                      data_collection_technique = "oth",
                      archived = "False")

p <- dataset_create(m)
#> Error in r$success: $ operator is invalid for atomic vectors
m <- resource_metadata(type = "data",
                       url = "mtcars.csv",
                       name = "mtcars.csv",
                       format = "csv",
                       file_type = "microdata",
                       date_range_start = "1973-01-01",
                       date_range_end = "1973-12-31",
                       version = "1",
                       process_status = "raw",
                       identifiability = "anonymized_public")
r <- resource_create(p$id, m)
#> Error in rlang::list2(...): object 'p' not found
# Like before, the return value is a tibble representation of the
# resource.

r
#> Error in eval(expr, envir, enclos): object 'r' not found

# But so far we’ve only created the metadata for the resource. The next
# step is to upload the data.
resource_upload(r$id, path = system.file("extdata/mtcars.csv", package = "readr"))
#> Error in resource_upload(r$id, path = system.file("extdata/mtcars.csv",     package = "readr")): could not find function "resource_upload"

resource_search("name:mtcars")
#> # A tibble: 23 × 31
#>    id           packa…¹ url   format descr…² hash  posit…³ name  resou…⁴ mimet…⁵
#>    <chr>        <chr>   <chr> <chr>  <chr>   <chr>   <int> <chr> <lgl>   <chr>  
#>  1 3acdcd5a-4f… eb71cf… http… CSV    "Some … ""          0 mtca… NA      applic…
#>  2 c70cdca6-83… 017591… http… CSV    ""      ""          0 mtca… NA      applic…
#>  3 a8c562fb-35… 5f01ae… http… TXT    "Some … ""          0 mtca… NA      text/p…
#>  4 aa464ce4-0d… 07b30a… http… CSV    "Some … ""          0 mtca… NA      NA     
#>  5 6f698bbe-77… 005277… http… CSV    "Some … ""          0 mtca… NA      applic…
#>  6 071b3fce-df… 08ffbb… http… CSV    "Some … ""          0 mtca… NA      applic…
#>  7 debc8207-ad… ac166d… http… CSV    "Some … ""          0 mtca… NA      applic…
#>  8 ea4aadda-13… 72de44… http… CSV    "Some … ""          0 mtca… NA      applic…
#>  9 63f1bdc6-b7… aa6eaa… http… CSV    "Some … ""          0 mtca… NA      NA     
#> 10 689cbcac-2c… eb71cf… http… CSV    "Some … ""          1 mtca… NA      applic…
#> # … with 13 more rows, 21 more variables: mimetype_inner <lgl>, size <int>,
#> #   created <chr>, last_modified <chr>, metadata_modified <chr>,
#> #   cache_url <lgl>, cache_last_updated <lgl>, url_type <chr>, state <chr>,
#> #   type <chr>, version <chr>, `hxl-ated` <chr>, file_type <chr>,
#> #   visibility <chr>, date_range_end <chr>, file_to_upload <chr>,
#> #   process_status <chr>, identifiability <chr>, datastore_active <lgl>,
#> #   date_range_start <chr>, title <chr>, and abbreviated variable names …