Work with RIDL resources (files)

Work with RIDL resources (files)

Usage

resource_create(pkgid, res_metadata)

resource_update(id, res_metadata)

resource_patch(id, res_metadata)

resource_delete(id)

resource_create(pkgid, res_metadata)

resource_update(id, res_metadata)

resource_patch(id, res_metadata)

resource_delete(id)

Arguments

pkgid

The id or name of the dataset to which this resource belongs to.

res_metadata

Metadata created by resource_metadata().

id

The id or name of the resource.

Value

The resource.

Details

You must have the necessary permissions to create, edit, or delete datasets and their resources.

Note that several fields are required for resource_create() and resource_update() operations to succeed. Consult resource_metadata() for the details.

For resource_update()/resource_patch() operations, it is recommended to call resource_show(), make the desired changes to the result, and then call resource_update()/resource_patch() with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the metadata.

Examples

## let's get again the details of the dataset we want to add the resource in..
#-----
# Test search in prod
Sys.unsetenv("USE_UAT")
p <- riddle::dataset_search("rms_v4")
p
#> # A tibble: 10 × 54
#>    admin_notes    archi…¹ author autho…² clean…³ creat…⁴ data_…⁵ data_…⁶ data_…⁷
#>    <chr>          <chr>   <lgl>  <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 "The RMS surv… False   NA     NA      ""      5cdbcb… ""      ""      tri    
#>  2 "R4V - RMRP B… True    NA     NA      ""      3b109e… "Restr… "Tool:… cawi   
#>  3 ""             False   NA     NA      ""      183804… ""      ""      tri    
#>  4 "Uploaded dat… False   NA     NA      "Uploa… 10a321… "Data … "Data … cami   
#>  5 ""             False   NA     NA      ""      de99be… ""      ""      tri    
#>  6 ""             True    NA     NA      ""      1b633b… ""      ""      tri    
#>  7  NA            False   NA     NA       NA     b5d85f…  NA      NA     tri    
#>  8  NA            False   NA     NA       NA     b5d85f…  NA      NA     tri    
#>  9 ""             False   NA     NA      ""      9a1c9d… ""      ""      capi   
#> 10 ""             False   NA     NA      ""      e44763… ""      ""      f2f    
#> # … with 45 more variables: data_collector <chr>, data_sensitivity <chr>,
#> #   date_range_end <chr>, date_range_start <chr>, external_access_level <chr>,
#> #   geog_coverage <chr>, geographies <list>, id <chr>, identifiability <chr>,
#> #   isopen <lgl>, keywords <list>, kobo_asset_id <chr>, license_id <lgl>,
#> #   license_title <lgl>, maintainer <lgl>, maintainer_email <lgl>,
#> #   metadata_created <chr>, metadata_modified <chr>, name <chr>, notes <chr>,
#> #   num_resources <int>, num_tags <int>, operational_purpose_of_data <list>, …

#-----
# Test search in uat
Sys.setenv(USE_UAT=1)
p <- riddle::dataset_search("tests")
p
#> # A tibble: 10 × 54
#>    admin_notes    archi…¹ author autho…² clean…³ creat…⁴ data_…⁵ data_…⁶ data_…⁷
#>    <chr>          <chr>   <lgl>  <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 ""             False   NA     NA      ""      daf9d9… ""      ""      nf     
#>  2 ""             True    NA     NA      ""      1b633b… ""      ""      tri    
#>  3 ""             False   NA     NA      ""      9a7869… ""      ""      tri    
#>  4 ""             False   NA     NA      ""      a87b60… ""      ""      f2f    
#>  5 ""             False   NA     NA      ""      f1b4d3… ""      ""      tri    
#>  6 "The dataset … False   NA     NA      "For p… 274891… ""      "The m… capi   
#>  7 ""             False   NA     NA      ""      aaf6de… ""      ""      tri    
#>  8 ""             True    NA     NA      "The i… cf0c93… "The K… "Eleve… tri    
#>  9 ""             True    NA     NA      "The i… cf0c93… "The K… "Six U… tri    
#> 10 ""             False   NA     NA      ""      f1b4d3… ""      ""      f2f    
#> # … with 45 more variables: data_collector <chr>, data_sensitivity <chr>,
#> #   date_range_end <chr>, date_range_start <chr>, external_access_level <chr>,
#> #   geog_coverage <chr>, geographies <list>, id <chr>, identifiability <chr>,
#> #   isopen <lgl>, keywords <list>, kobo_asset_id <chr>, license_id <lgl>,
#> #   license_title <lgl>, maintainer <lgl>, maintainer_email <lgl>,
#> #   metadata_created <chr>, metadata_modified <chr>, name <chr>, notes <chr>,
#> #   num_resources <int>, num_tags <int>, operational_purpose_of_data <list>, …

ridlid <- as.character(p[1, c("id")])

m <- riddle::resource_metadata(type = "data",
                       url = "mtcars.csv",
                       name = "mtcarsriddle",
                       format = "csv",
                       file_type = "microdata",
                       date_range_start = "1973-01-01",
                       date_range_end = "1973-12-31",
                       version = "1",
                       visibility = "public",
                       process_status = "raw",
                       identifiability = "anonymized_public",
upload = httr::upload_file(system.file("extdata/mtcars.csv", package = "readr")))

## let's get again the details of the dataset we want to add the resource in..
#r <- riddle::resource_create(pkgid = ridlid, res_metadata = m)
# Like before, the return value is a tibble representation of the resource.
#r
 
## and now can search for it - checking it is correctly there... 
##riddle::resource_search("name:mtcarsriddle")

# And once we’re done experimenting with the API, we should take down our
# toy dataset since we don’t really need it on RIDL.
#riddle::dataset_delete(p$id)